begin en end pos working for normalized element start tags that means no extra spaces in the tag, also no self closing start tag
